feat(search): add quiescence search and move ordering
- Implements `quiescence()` to mitigate the horizon effect by searching
  all captures until a quiet position is reached.
- Integrates `score_list()` and `next_move()` in the search loop to
  improve alpha-beta pruning efficiency.
- Updates Piece-Square Tables (PSQT) with tuned values for both middle
  and end games.
- Refines FEN parsing to handle empty castling rights and missing
  half-move clocks.
- Adjusts UCI time control logic and move overhead calculations.
- Removes legacy `main.c` and changes `score_t` components to 32-bit
  integers for better alignment/performance.
- Ensures stdout is flushed after sending bestmove.

+++ src/psqt.h
@@ -7,8 +7,8 @@
 #define TOTAL_PHASE 24  // Phase at starting position
 
 typedef struct {
-  int16_t mg;
-  int16_t eg;
+  int mg;
+  int eg;
 } score_t;
 
 extern const score_t PSQT[NR_OF_PIECE_TYPES][NR_OF_SQUARES];
